+++ kbox2d/src/commonMain/kotlin/org/jbox2d/dynamics/contacts/Position.kt
@@ -29,5 +29,5 @@ class Position {
 
     val c = Vec2()
 
-    var a: Float = 0f
+    var a: Float = 0.toFloat()
 }
